+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update video title when refreshing if there is a change.
 - Show an indicator next to members only videos.
 - Separate videos, streams and shorts tabs.
+- Add an option for original audio track preference
 
 ### Changed
 - Reduce `tick_rate` default value to 10 milliseconds.